requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.5
cocos2d==0.6.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms==1.12.0
django-excel==0.0.10
django-heroku==0.3.1
django-import-export==2.5.0
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
lml==0.1.0
MarkupPy==1.14
mysqlclient==2.0.3
newrelic==6.10.0.165
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.2
platformdirs==2.2.0
psycopg2==2.9.1
pyexcel==0.6.7
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pygame==2.0.1
pyglet==1.5.18
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
texttable==1.6.4
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.7.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
